requirements.txt
altair==5.5.0
appnope==0.1.4
asttokens==2.4.1
attrs==25.3.0
Backtesting==0.3.3
beautifulsoup4==4.12.3
blinker==1.9.0
bokeh==3.4.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.3.2
click==8.1.8
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
filelock==3.16.1
fonttools==4.51.0
fredapi==0.5.2
frozendict==2.4.4
fsspec==2024.9.0
ftfy==6.3.0
gitdb==4.0.12
GitPython==3.1.44
googlemaps==4.10.0
html5lib==1.1
huggingface-hub==0.25.2
idna==3.7
introcs==1.3.1
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lxml==5.2.2
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mpmath==1.3.0
multitasking==0.0.11
narwhals==1.35.0
nest-asyncio==1.6.0
networkx==3.4.1
numpy==1.26.2
openai-clip==1.0.1
packaging==24.0
pandas==2.2.2
pandas-datareader==0.10.0
parso==0.8.4
peewee==3.17.5
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.2.0
plotly==6.0.1
praw==7.8.1
prawcore==2.4.0
prompt-toolkit==3.0.43
protobuf==5.29.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==19.0.1
pydeck==0.9.1
Pygments==2.17.2
pyparsing==3.1.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.2
referencing==0.36.2
regex==2024.9.11
requests==2.32.2
rpds-py==0.24.0
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==69.0.2
six==1.16.0
smmap==5.0.2
soupsieve==2.5
stack-data==0.6.3
streamlit==1.44.1
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.20.1
toml==0.10.2
torch==2.4.1
torchvision==0.19.1
tornado==6.4
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.2
typing_extensions==4.12.2
tzdata==2024.1
update-checker==0.18.0
urllib3==2.2.1
vaderSentiment==3.3.2
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
whatever==0.7
wheel==0.42.0
wordcloud==1.9.4
xyzservices==2024.4.0
yfinance==0.2.54
